쉴드와 무기 변형만 남음

+++ Parkjein/Assets/_Scripts/HanSocket/VO/InGame/DamageVO.cs
@@ -19,6 +19,6 @@ namespace HanSocket.VO.InGame
     public class SpecialCommands
     {
         public string command;
-        public int param;
+        public float param;
     }
 }
